+++ Sourcecode/Song.Entities/Examination.cs
@@ -38,7 +38,7 @@ namespace Song.Entities {
     		
     		protected String _Th_Name;
     		
-    		protected Int32 _Tp_Id;
+    		protected Int64 _Tp_Id;
     		
     		protected Int32 _Exam_PassScore;
     		
@@ -260,7 +260,7 @@ namespace Song.Entities {
     		/// <summary>
     		/// -1
     		/// </summary>
-    		public Int32 Tp_Id {
+    		public Int64 Tp_Id {
     			get {
     				return this._Tp_Id;
     			}
@@ -523,7 +523,7 @@ namespace Song.Entities {
     				this._Th_Name = reader.GetString(_.Th_Name);
     			}
     			if ((false == reader.IsDBNull(_.Tp_Id))) {
-    				this._Tp_Id = reader.GetInt32(_.Tp_Id);
+    				this._Tp_Id = reader.GetInt64(_.Tp_Id);
     			}
     			if ((false == reader.IsDBNull(_.Exam_PassScore))) {
     				this._Exam_PassScore = reader.GetInt32(_.Exam_PassScore);
@@ -660,7 +660,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Th_Name = new WeiSha.Data.Field<Examination>("Th_Name");
     			
     			/// <summary>
-    			/// -1 - 字段名：Tp_Id - 数据类型：Int32
+    			/// -1 - 字段名：Tp_Id - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Tp_Id = new WeiSha.Data.Field<Examination>("Tp_Id");
